Improved code for restoring light colors after canceling

+++ javahueapi/src/main/java/nl/q42/javahueapi/models/State.java
@@ -15,7 +15,7 @@ public class State implements Serializable {
 	
 	public int sat;
 	
-	public double[] xy;
+	public float[] xy;
 	
 	public int ct;
